first fix

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImp.java
@@ -1,10 +1,8 @@
 package ru.kata.spring.boot_security.demo.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.kata.spring.boot_security.demo.dao.UserDao;
@@ -14,11 +12,8 @@ import java.util.List;
 
 @Service
 public class UserServiceImp implements UserDetailsService, UserService {
-
-   private UserDao userDao;
-
-   @Autowired
-   public void setUserDao(UserDao userDao) {
+   private final UserDao userDao;
+   public  UserServiceImp(UserDao userDao) {
       this.userDao = userDao;
    }
 
@@ -31,6 +26,9 @@ public class UserServiceImp implements UserDetailsService, UserService {
    @Transactional
    @Override
    public void delete(User user) {
+      if (!userDao.isUserExists(user)) {
+         throw new IllegalArgumentException("User with id " + user.getId() + " not found");
+      }
       userDao.delete(user);
    }
